Move context auth provider from @crossmint/client-sdk-auth-core to @crossmint/client-sdk-react-ui
