--- packages/client/auth-core/client/index.ts
@@ -1,1 +1,0 @@
-export { useAuth, AuthProvider, type AuthProviderParams, getCachedJwt } from "../src/contexts/authlogin";
