--- packages/client/auth-core/client/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "jsx": "react-jsx",
-        "target": "es6"
-    }
-}
