+++ packages/client/auth-core/package.json
@@ -7,10 +7,6 @@
     "sideEffects": false,
     "type": "module",
     "exports": {
-        "./client": {
-            "import": "./dist/client/index.js",
-            "require": "./dist/client/index.cjs"
-        },
         "./server": {
             "import": "./dist/server/index.js",
             "require": "./dist/server/index.cjs"
@@ -22,28 +18,16 @@
         "LICENSE"
     ],
     "scripts": {
-        "build": "tsup client/index.ts server/index.ts --clean --external react,react-dom --format esm,cjs --outDir ./dist --minify --dts --sourcemap",
-        "dev": "tsup client/index.ts server/index.ts --clean --external react,react-dom --format esm,cjs --outDir ./dist --dts --sourcemap --watch",
+        "build": "tsup server/index.ts --clean --format esm,cjs --outDir ./dist --minify --dts --sourcemap",
+        "dev": "tsup server/index.ts --clean --format esm,cjs --outDir ./dist --dts --sourcemap --watch",
         "test": "jest --passWithNoTests"
     },
     "dependencies": {
         "@crossmint/client-sdk-base": "workspace:*",
-        "@crossmint/client-sdk-window": "workspace:*",
-        "@crossmint/common-sdk-base": "workspace:*",
-        "@headlessui/react": "1.7.18",
         "jsonwebtoken": "9.0.2",
-        "jwks-rsa": "3.1.0",
-        "zod": "3.22.4"
+        "jwks-rsa": "3.1.0"
     },
     "devDependencies": {
-        "@types/jsonwebtoken": "9.0.6",
-        "@types/react": "^18.3.0",
-        "@types/react-dom": "^18.3.0",
-        "react": "^18.3.0",
-        "react-dom": "^18.3.0"
-    },
-    "peerDependencies": {
-        "react": ">=17.0.2",
-        "react-dom": ">=17.0.2"
+        "@types/jsonwebtoken": "9.0.6"
     }
 }
