+++ packages/client/ui/react-ui/package.json
@@ -30,14 +30,22 @@
         "@crossmint/client-sdk-auth-core": "workspace:*",
         "@crossmint/client-sdk-base": "workspace:*",
         "@crossmint/client-sdk-smart-wallet": "workspace:*",
+        "@crossmint/client-sdk-window": "workspace:*",
         "@crossmint/common-sdk-base": "workspace:*",
         "@ethersproject/transactions": "5.7.0",
+        "@headlessui/react": "2.1.5",
+        "@radix-ui/react-dialog": "1.1.1",
         "@solana/web3.js": "1.95.1",
         "bs58": "5.0.0",
+        "jsonwebtoken": "9.0.2",
+        "jwks-rsa": "3.1.0",
         "lodash.isequal": "4.5.0",
-        "react-jss": "10.10.0"
+        "react-jss": "10.10.0",
+        "tailwindcss": "3.4.10",
+        "zod": "3.22.4"
     },
     "devDependencies": {
+        "@types/jsonwebtoken": "9.0.6",
         "@types/lodash.isequal": "4.5.6",
         "@types/react": "^18.3.0",
         "@types/react-dom": "^18.3.0",
