+++ packages/client/ui/react-ui/src/components/index.ts
@@ -3,3 +3,4 @@ export * from "./CrossmintNFTDetail";
 
 export * from "./embed";
 export * from "./hosted";
+export * from "./auth";
