+++ packages/client/ui/react-ui/src/hooks/index.ts
@@ -1,3 +1,3 @@
 export * from "./useCrossmint";
 export * from "./useWallet";
-export { useAuth } from "@crossmint/client-sdk-auth-core/client";
+export * from "./useAuth";
