+++ packages/client/ui/react-ui/src/hooks/useAuth.ts
@@ -1,0 +1,10 @@
+import { AuthContext } from "@/providers";
+import { useContext } from "react";
+
+export function useAuth() {
+    const context = useContext(AuthContext);
+    if (context === undefined) {
+        throw new Error("useAuth must be used within an AuthProvider");
+    }
+    return context;
+}
