+++ packages/client/ui/react-ui/src/hooks/useCrossmint.tsx
@@ -1,6 +1,6 @@
+import { getCachedJwt } from "@/utils";
 import { ReactNode, createContext, useCallback, useContext, useMemo, useRef, useState } from "react";
 
-import { getCachedJwt } from "@crossmint/client-sdk-auth-core/client";
 import { Crossmint, createCrossmint } from "@crossmint/common-sdk-base";
 
 export interface CrossmintContext {
