+++ packages/client/ui/react-ui/src/services/logger.ts
@@ -1,0 +1,5 @@
+import { AUTH_SERVICE } from "@/utils";
+
+import { SDKLogger } from "@crossmint/client-sdk-base";
+
+export const authLogger = new SDKLogger(AUTH_SERVICE);
